requirements.txt
# Core web framework
streamlit==1.29.0

# PDF/Document parsing
PyMuPDF==1.23.8
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23
pdf2image==1.16.3
Pillow==10.1.0

# OCR
pytesseract==0.3.10
# pix2tex (optional, install separately: pip install pix2tex)

# Embeddings and Vector DB
sentence-transformers==2.2.2
chromadb==0.4.22
faiss-cpu==1.8.0

# RAG and LLM
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.6.1

# Transformers (for local LLM fallback)
transformers==4.36.2
accelerate==0.25.0
bitsandbytes==0.41.3
torch==2.2.0

# Validators and utilities
sympy==1.12
scikit-learn==1.3.2
numpy==1.26.2
scipy==1.11.4
tiktoken==0.5.2

# API and utilities
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# PDF Generation
reportlab==4.0.7

# Logging and monitoring
python-dotenv==1.0.0
